Dev/4.6.0

+++ .github/ci/build/build_android.sh
@@ -37,6 +37,7 @@
 # pr: output test.zip to workspace dir
 # others: Rename the zip package name yourself, But need copy it to workspace dir
 ##################################
+export PATH=$PATH:/opt/homebrew/bin
 
 echo Package_Publish: $Package_Publish
 echo is_tag_fetch: $is_tag_fetch
@@ -49,6 +50,8 @@ echo release_version: $release_version
 echo short_version: $short_version
 echo pwd: `pwd`
 echo sdk_url: $sdk_url
+echo android_direction: $android_direction
+
 unzip_name=Agora_Native_SDK_for_Android_FULL_DEFAULT
 zip_name=Agora_Native_SDK_for_Android_FULL_DEFAULT.zip
 if [ -z "$sdk_url" ] || [ "$sdk_url" = "none" ]; then
@@ -63,33 +66,46 @@ else
 	curl -o $zip_name $sdk_url || exit 1
 	7za x ./$zip_name -y > log.txt
 
-	unzip_name=`ls -S -d */ | grep Agora | sed 's/\///g'`
+	# Support top-level directory name containing 'Agora' or 'Shengwang'
+	unzip_name=`ls -S -d */ | grep -E 'Agora|Shengwang' | head -n 1 | sed 's/\///g'`
+	if [ -z "$unzip_name" ]; then
+		echo "Error: Unzipped directory not found. The SDK package structure may be invalid or the top-level directory does not contain 'Agora' or 'Shengwang'"
+		exit 1
+	fi
 	echo unzip_name: $unzip_name
 
 	rm -rf ./$unzip_name/rtc/bin
 	rm -rf ./$unzip_name/rtc/demo
-	rm ./$unzip_name/rtc/commits
-	rm ./$unzip_name/rtc/package_size_report.txt
+	rm -f ./$unzip_name/.commits
+	rm -f ./$unzip_name/spec
 	rm -rf ./$unzip_name/pom
 fi
-mkdir -p ./$unzip_name/rtc/samples 
-cp -rf ./Android/${android_direction} ./$unzip_name/rtc/samples/API-Example || exit 1
+
+mkdir -p ./$unzip_name/rtc/samples/${android_direction} || exit 1
+rm -rf ./$unzip_name/rtc/samples/${android_direction}/*
+
+if [ -d "./Android/${android_direction}" ]; then
+    cp -rf ./Android/${android_direction}/* ./$unzip_name/rtc/samples/${android_direction}/ || exit 1
+else
+    echo "Error: Source directory ./Android/${android_direction} does not exist"
+    exit 1
+fi
+
 7za a -tzip result.zip -r $unzip_name > log.txt
 mv result.zip $WORKSPACE/withAPIExample_${BUILD_NUMBER}_$zip_name
 
+if [ $compress_apiexample = true ]; then
+	onlyCodeZipName=${android_direction}_onlyCode.zip
+	7za a -tzip $onlyCodeZipName -r ./$unzip_name/rtc/samples/${android_direction} >> log.txt
+	mv $onlyCodeZipName $WORKSPACE/APIExample_onlyCode_${BUILD_NUMBER}_$zip_name
+fi
+
 if [ $compile_project = true ]; then
-	# install android sdk
-	which java
-	java --version
-	source ~/.bashrc
-	export ANDROID_HOME=/usr/lib/android_sdk
-	echo ANDROID_HOME: $ANDROID_HOME
-	cd ./$unzip_name/rtc/samples/API-Example || exit 1
+	cd ./$unzip_name/rtc/samples/${android_direction} || exit 1
 	if [ -z "$sdk_url" ] || [ "$sdk_url" = "none" ]; then
 		./cloud_build.sh false || exit 1
 	else
 		./cloud_build.sh true || exit 1
 	fi
 fi
-
